backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
openai
python-multipart